add chart page

+++ server/controllers/ProductController.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product')
+const gcsDelete = require('../helpers/gcsDelete')
 
 class ProductController {
   static create(req, res, next) {
@@ -35,11 +36,22 @@ class ProductController {
   static updateField(req, res, next) {
     let { id } = req.params
     let { name, price, stock, image } = req.body
-    if (!name || !price || !stock || !image) {
-      res.status(400).json({ message: 'bad request' })
+    if(req.file) {
+      Product.findById({ _id: id })
+        .then(product => {
+          gcsDelete(product.image)
+          return Product.findByIdAndUpdate({ _id: id }, { $set: { name, price: Number(price), stock, image } }, { runValidators: true, omitUndefined: true, new: true })
+        })
+        .then(result => {
+          res.status(200).json(result)
+        })
+        .catch(next)
     } else {
-      Product.findByIdAndUpdate({ _id: id }, { $set: { name, price: Number(price), stock, image } }, { runValidators: true, omitUndefined: true, new: true })
-      
+      Product.findById({ _id: id })
+        .select('image')
+        .then(product => {
+          return Product.findByIdAndUpdate({ _id: id }, { $set: { name, price: Number(price), stock, image: product.image } }, { runValidators: true, omitUndefined: true, new: true })
+        })
         .then(result => {
           res.status(200).json(result)
         })
@@ -49,7 +61,15 @@ class ProductController {
 
   static deleteProduct(req, res, next) {
     let { id } = req.params
-    Product.deleteOne({ _id: id })
+    Product.findById({ _id: id })
+      .then(result => {
+        if (result) {
+          gcsDelete(result.image)
+          return Product.deleteOne({ _id: id })
+        } else {
+          throw ({ status: 404, message: 'Product not found' })
+        }
+      })
       .then(result => {
         res.status(200).json(result)
       })
